moved search command
